perf: FPGA II=1 pipeline (246.3M TPS) + AVX2 SIMD hot path + 2-way cancel map
FPGA Matching Engine:
- Rewrote from 5-state blocking FSM (II=3) to 2-stage pipeline (II=1)
- Combinational BBO forwarding eliminates RAW hazard on consecutive inserts
- 246.3M orders/sec pipelined (98.5% of 250MHz physical limit)
- 99.1M orders/sec sustained crossing with fill generation
- 492.6M orders/sec projected @ 500MHz (Versal fast path)
- Fixed s2_sweep_depth bit-slice bug: fills now 1000/1000 correct
- Cancel map upgraded to 2-way set-associative (prevents hash collisions)
- All 8 simulation tests pass, 0 errors

Software (C++ AVX2/AVX-512):
- compiler_hints.hpp: SIMD dispatch library with target attributes
- orderbook.hpp: 4-way AVX2 hash probe (_mm256_cmpeq_epi64)
- ring_buffer.hpp: L1 prefetch on SPSC read path
- shm_engine.hpp: non-temporal streaming stores
- CompilerFlags.cmake: -O3 -flto -march=native -funroll-loops
- Cross-core SPSC improved +14.3% (LTO/PLT elimination)

Docs updated: README, ARCHITECTURE_BOOK, FPGA_SIMULATION_REPORT, BENCHMARK_METHODOLOGY

+++ core/include/core/hdr_histogram.hpp
@@ -42,7 +42,7 @@ private:
     uint64_t total_count_ = 0;
     uint64_t min_value_ = UINT64_MAX;
     uint64_t max_value_ = 0;
-    double total_sum_ = 0.0;
+    uint64_t total_sum_ = 0;      // Integer sum — cast to double only in get_percentiles()
     uint64_t lowest_trackable_ = 1;
     uint64_t highest_trackable_ = 10'000'000'000ULL;
     int sub_bucket_half_count_magnitude_ = 0;
